requirements.txt
langchain
langchain-openai
langchain-core
langchain-text-splitters
langchain-community
langchain-chroma
chromadb
fastapi
uvicorn[standard]
python-dotenv
unstructured[image]
httpx==0.27.2
beautifulsoup4==4.12.3
pypdf==4.3.1
python-docx==1.1.2
rarfile
